copy begs waybar config
this has already been modified from the source to remove the powerline dividers and remove some of the modules. Font moved to Cozette bitmap font which looks good and has a few icons as well. further changes to the config still need to be made to make use of colours from dotdrop. gammastep has also been removed in favor of wlsunset which is easier to work with